requirements.txt
langchain
langchain-core
langchain-google-vertexai
chromadb
datasets
ragas
google-cloud-aiplatform
google-auth
unstructured
unidecode
markdownify
unstructured[md]